Fixes in distributed tcp/sessions storage
git-svn-id: svn://svn.code.sf.net/p/cppcms/code/framework@606 831565cc-8e21-4ad5-8334-98019c97bd36

+++ branches/sessions/session_sqlite_storage.cpp
@@ -83,7 +83,7 @@ public:
 	}
 	void check_commits()
 	{
-		int64_t now=time(NULL);
+		long long int now=time(NULL);
 		if(	write_ops >= deferred_commit_count 
 			|| (deferred_commit_time > 0 && last_commit + deferred_commit_time < now))
 		{
